+++ jdempotent/src/test/java/br/com/sawcunhaos/foundation/jdempotent/core/aspect/TestAopTransactionalContext.java
@@ -37,7 +37,10 @@ class TestAopTransactionalContext {
 
     @Bean
     IdempotentAspect idempotentAspect(InMemoryIdempotentRepository inMemoryIdempotentRepository, DefaultKeyGenerator defaultKeyGenerator) {
-        return new IdempotentAspect(inMemoryIdempotentRepository, defaultKeyGenerator);
+        return IdempotentAspect.builder()
+                .repository(inMemoryIdempotentRepository)
+                .keyGenerator(defaultKeyGenerator)
+                .build();
     }
 
     @Bean